+++ src/defaultsettings.cpp
@@ -83,6 +83,7 @@ void fm_set_default_settings(Settings *settings) {
 	settings->setDefault("screenshot_path", "screenshots"); // "."
 	settings->setDefault("serverlist_file", "favoriteservers.json"); // "favoriteservers.txt"
 	settings->setDefault("serverlist_cache", porting::path_user + DIR_DELIM + "client" + DIR_DELIM + "servers_public.json");
+	settings->setDefault("serverlist_lan", "1");
 
 	// Main menu
 	settings->setDefault("main_menu_tab", "multiplayer");